+++ lambda-function/contract/abi.js
@@ -1,9 +1,9 @@
-import MolochABI from "./abi/moloch-abi.json" assert { type: "json" };
-import erc20ABI from "./abi/erc20.json" assert { type: "json" };
+const MolochABI = require("./abi/moloch-abi.json");
+const erc20ABI = require("./abi/erc20.json");
 
 const abi = {
   erc20ABI,
   MolochABI,
 };
 
-export default abi;
+module.exports = abi;
