+++ lambda-function/contract/index.js
@@ -1,6 +1,3 @@
-// import abi from "./abi.js";
-// import address from "./address.js";
-
 const abi = require("./abi.js");
 const address = require("./address.js");
