+++ lambda-function/mutations.js
@@ -1,8 +1,6 @@
-// import axios from "axios";
-// convert above to require
 const axios = require("axios");
 
-export async function insertTokenValue(tokens_to_hasura) {
+async function insertTokenValue(tokens_to_hasura) {
   const query = `
     mutation insertTokenInfo($tokens_to_hasura: [treasury_token_history_insert_input!]!) {
       insert_treasury_token_history(objects: $tokens_to_hasura) {
@@ -31,3 +29,5 @@ export async function insertTokenValue(tokens_to_hasura) {
   // }, { headers });
   return response;
 }
+
+module.exports = { insertTokenValue };
